# Speed up import time by deferring inspect