@jaraco 2024-08-20
I've minimized the change to address the specific issue. I'll deal with the typeshed ignore workaround separately.