@jaraco 2024-08-20
I should have added a news fragment before merging. I added it later in 71b4678432.