@jaraco 2024-08-20
This change is released in v8.4.0.